requirements.txt
streamlit>=1.28.0
requests>=2.31.0
google-generativeai>=0.3.0

# Document processing (optional)
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Image processing (optional)
Pillow>=10.0.0

# Web search (optional)
duckduckgo-search>=3.9.0

#setup tools
setuptools>=67.0.0